requirements.txt
hydra-core==1.3.2
imageio==2.31.1
matplotlib==3.7.1
numpy==1.25.0
omegaconf==2.3.0
pandas==2.0.2
Pillow==9.5.0
pytorch_lightning==1.6.5
ray==2.5.1
ray_lightning==0.3.0
scikit_learn==1.2.2
setuptools==68.0.0
torch_geometric==2.3.1
torchmetrics==0.11.4
tqdm==4.65.0
wandb==0.15.4